+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/atomist-sdm/compare/0.1.1...HEAD
 
+### Added
+
+-   Publish TypeDoc when Node project is released
+-   Increment version after release
+
 ## [0.1.1][] - 2018-05-10
 
 [0.1.1]: https://github.com/atomist/atomist-sdm/compare/0.1.0...0.1.1